Refactor engine structure and add UI sandbox
