+++ CMakeLists.txt
@@ -1,13 +1,12 @@
 cmake_minimum_required(VERSION 3.25)
-project(WidgeCraft VERSION 0.2.0 LANGUAGES CXX)
+project(WidgeCraft VERSION 0.3.0 LANGUAGES CXX)
 
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
+set_property(GLOBAL PROPERTY USE_FOLDERS ON)
 
-option(WIDGECRAFT_BUILD_DEMO "Build the interactive WidgeCraft showcase" ON)
-option(WIDGECRAFT_BUILD_SANDBOX "Build the interactive 3D raycast sandbox" ON)
-option(WIDGECRAFT_BUILD_TESTS "Build WidgeCraft unit tests" ON)
+option(WIDGECRAFT_BUILD_SANDBOX "Build the WidgeCraft UI sandbox" ON)
 
 include(FetchContent)
 
@@ -43,23 +42,55 @@ function(widgecraft_enable_warnings target)
     endif()
 endfunction()
 
-# ---------------- Engine library ----------------
-add_library(widgecraft_engine STATIC
-    src/Window.cpp
-    src/Input.cpp
-    src/TextRenderer.cpp
-    src/ShapeRenderer.cpp
-    src/Frame.cpp
-    src/Widget.cpp
-    src/Raycast.cpp
-    src/WidgeCraft.cpp
+set(WIDGECRAFT_ENGINE_HEADERS
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/WidgeCraft/WidgeCraft.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/WidgeCraft/input/Input.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/WidgeCraft/window/Window.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/WidgeCraft/ui/Frame.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/WidgeCraft/ui/Widget.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/WidgeCraft/scene/Scene.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/WidgeCraft/scene/Raycast.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/WidgeCraft/render/ShaderProgram.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/WidgeCraft/primitives/Types.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/WidgeCraft/primitives/Shapes2D.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/WidgeCraft/primitives/Shapes3D.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/WidgeCraft/primitives/TextRenderer.hpp
 )
 
+set(WIDGECRAFT_ENGINE_SOURCES
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/core/WidgeCraft.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/input/Input.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/window/Window.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/ui/Frame.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/ui/Widget.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/scene/Raycast.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/render/ShaderProgram.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/primitives/Shapes2D.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/primitives/Shapes3D.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/primitives/TextRenderer.cpp
+)
+
+# ---------------- Engine library ----------------
+add_library(widgecraft_engine STATIC
+    ${WIDGECRAFT_ENGINE_HEADERS}
+    ${WIDGECRAFT_ENGINE_SOURCES}
+)
 add_library(WidgeCraft::Engine ALIAS widgecraft_engine)
 
 set_target_properties(widgecraft_engine PROPERTIES
     OUTPUT_NAME WidgeCraft
-    FOLDER "WidgeCraft"
+    FOLDER "Engine"
+)
+
+source_group(
+    TREE ${CMAKE_CURRENT_SOURCE_DIR}/include
+    PREFIX "Header Files"
+    FILES ${WIDGECRAFT_ENGINE_HEADERS}
+)
+source_group(
+    TREE ${CMAKE_CURRENT_SOURCE_DIR}/src
+    PREFIX "Source Files"
+    FILES ${WIDGECRAFT_ENGINE_SOURCES}
 )
 
 target_include_directories(widgecraft_engine
@@ -77,54 +108,26 @@ target_compile_definitions(widgecraft_engine
 target_link_libraries(widgecraft_engine PUBLIC glfw glad)
 widgecraft_enable_warnings(widgecraft_engine)
 
-# ---------------- Interactive showcase ----------------
-if(WIDGECRAFT_BUILD_DEMO)
-    add_executable(widgecraft src/main.cpp)
-    set_target_properties(widgecraft PROPERTIES FOLDER "Examples")
-    target_link_libraries(widgecraft PRIVATE WidgeCraft::Engine)
-    widgecraft_enable_warnings(widgecraft)
-
-    add_custom_command(TARGET widgecraft POST_BUILD
-        COMMAND ${CMAKE_COMMAND} -E copy_directory
-            ${CMAKE_CURRENT_SOURCE_DIR}/assets
-            $<TARGET_FILE_DIR:widgecraft>/assets
-        COMMENT "Copying WidgeCraft assets"
-    )
-endif()
-
-# ---------------- Interactive raycast sandbox test ----------------
+# ---------------- Launchable sandboxes ----------------
 if(WIDGECRAFT_BUILD_SANDBOX)
-    add_executable(widgecraft_3d_sandbox tests/ThreeDSandbox.cpp)
-    set_target_properties(widgecraft_3d_sandbox PROPERTIES FOLDER "Tests")
-    target_link_libraries(widgecraft_3d_sandbox PRIVATE WidgeCraft::Engine)
-    widgecraft_enable_warnings(widgecraft_3d_sandbox)
+    add_executable(widgecraft_ui_sandbox
+        ${CMAKE_CURRENT_SOURCE_DIR}/sandbox/UiDemo.cpp
+    )
+    set_target_properties(widgecraft_ui_sandbox PROPERTIES
+        FOLDER "Sandbox"
+    )
+    target_link_libraries(widgecraft_ui_sandbox PRIVATE WidgeCraft::Engine)
+    widgecraft_enable_warnings(widgecraft_ui_sandbox)
 
-    # Make the generated Visual Studio solution launch the sandbox by default.
     set_property(
         DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
-        PROPERTY VS_STARTUP_PROJECT widgecraft_3d_sandbox
+        PROPERTY VS_STARTUP_PROJECT widgecraft_ui_sandbox
     )
 
-    add_custom_command(TARGET widgecraft_3d_sandbox POST_BUILD
+    add_custom_command(TARGET widgecraft_ui_sandbox POST_BUILD
         COMMAND ${CMAKE_COMMAND} -E copy_directory
             ${CMAKE_CURRENT_SOURCE_DIR}/assets
-            $<TARGET_FILE_DIR:widgecraft_3d_sandbox>/assets
-        COMMENT "Copying WidgeCraft assets for 3D sandbox"
+            $<TARGET_FILE_DIR:widgecraft_ui_sandbox>/assets
+        COMMENT "Copying WidgeCraft assets for UI sandbox"
     )
-endif()
-
-# ---------------- Tests ----------------
-if(WIDGECRAFT_BUILD_TESTS)
-    include(CTest)
-    enable_testing()
-
-    add_executable(widgecraft_raycast_tests
-        tests/RaycastTests.cpp
-        src/Raycast.cpp
-    )
-    set_target_properties(widgecraft_raycast_tests PROPERTIES FOLDER "Tests")
-    target_include_directories(widgecraft_raycast_tests PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include)
-    widgecraft_enable_warnings(widgecraft_raycast_tests)
-
-    add_test(NAME WidgeCraft.Raycast COMMAND widgecraft_raycast_tests)
 endif()
