+++ generate_vs2022_win32.bat
@@ -3,7 +3,7 @@ setlocal
 
 where cmake >nul 2>nul
 if errorlevel 1 (
-    echo CMake was not found. Install CMake or run this from a Visual Studio Developer Command Prompt.
+    echo CMake was not found. Install the Visual Studio CMake tools first.
     pause
     exit /b 1
 )
@@ -21,7 +21,11 @@ if errorlevel 1 (
 set "SOLUTION=%CD%\build\win32-release\WidgeCraft.sln"
 echo.
 echo Generated: %SOLUTION%
+echo Startup project: widgecraft_ui_sandbox
 
-if exist "%SOLUTION%" start "" "%SOLUTION%"
+if /I not "%~1"=="--no-open" (
+    if exist "%SOLUTION%" start "" "%SOLUTION%"
+)
+
 popd
 endlocal
