+++ src/primitives/TextRenderer.cpp
