--- tests/ThreeDSandbox.cpp
@@ -1,4 +1,0 @@
-// Interactive raycast integration-test entry point.
-// Kept beside RaycastTests.cpp so the raycast tests and visual sandbox
-// are grouped together in Visual Studio.
-#include "../examples/ThreeDSandbox.cpp"
